Support imagePullSecrets for sonobuoy workers
In the situation where the sonobuoy image itself was
in a private registry there was no recourse for users
because imagePullSecrets was not defined in the
config object used to help populate the templates
and these templates were generated on the server
side (so the user couldn't use sonobuoy gen and
manually manipulate them).

This change does the following:
 - adds imagePullSecrets to the config type
 - adds sections in the templates to honor it

It does not modify the original template to
add the secret or apply it to the aggregator
pod, those steps must be done manually after
sonobuoy gen.

This at least unblocks people who may be in
this situation.

Fixes #424